requirements.txt
streamlit
transformers
accelerate
pdfplumber
python-dotenv
